shooting bird

+++ shootingrobot/src/com/smp/shootingrobot/Grass.java
@@ -2,7 +2,7 @@ package com.smp.shootingrobot;
 
 public class Grass extends Scrollable {
 
-	public Grass(float x, float y, int width, int height, float scrollSpeed) {
+	public Grass(float x, float y, float width, float height, float scrollSpeed) {
 		super(x, y, width, height, scrollSpeed);
 	}
